ai/requirements.txt
accelerate
firebase-admin==6.5.0
huggingface_hub[cli]
hydra-core
jsonlines
matplotlib
numpy==1.25.2
openai
pandas==2.1.4
pillow==10.2.0
pydantic
pyyaml
scikit-learn==1.3.0
scipy==1.11.4
shapely==2.0.1
supervision==0.22.0
tqdm
transformers==4.44.0
